feat(servicenow): reuse frontend data type
