+++ workspaces/servicenow/plugins/servicenow-backend/src/service/router.ts
@@ -27,9 +27,13 @@ import {
 import express from 'express';
 import Router from 'express-promise-router';
 import { ServiceNowSingleConfig } from '../config';
-import { DefaultServiceNowClient } from '../service-now-rest/client';
+import {
+  DefaultServiceNowClient,
+  IncidentPick,
+} from '../service-now-rest/client';
 import type { CatalogApi } from '@backstage/catalog-client';
 import { Entity } from '@backstage/catalog-model';
+import { IncidentsData } from '@backstage-community/plugin-servicenow-common';
 
 export interface RouterOptions {
   logger: LoggerService;
@@ -153,7 +157,7 @@ export async function createRouter(
       }
 
       const incidents = await client.fetchIncidents(fetchOptions);
-      res.json(incidents);
+      res.json(incidentsPickToIncidentsData(incidents));
     } catch (error) {
       if (error instanceof InputError) {
         throw error;
@@ -166,4 +170,15 @@ export async function createRouter(
   });
 
   return router;
+}
+
+function incidentsPickToIncidentsData(data: IncidentPick[]): IncidentsData[] {
+  return data.map(item => ({
+    number: item.number,
+    shortDescription: item.short_description,
+    description: item.description,
+    sysCreatedOn: item.sys_created_on,
+    priority: item.priority,
+    incidentState: item.incident_state,
+  }));
 }
