+++ workspaces/servicenow/plugins/servicenow/src/types.ts
@@ -13,14 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export type IncidentsData = {
-  number: string;
-  shortDescription: string;
-  description: string;
-  sysCreatedOn: string;
-  priority: number;
-  incidentState: number;
-};
 
 export enum IncidentTableFieldEnum {
   Number = 'number',
